llcppsymg:pkg parse -> symbol/parse.go
