--- _xtool/llcppsymg/_cmptest/parse_test/llgo.expect
@@ -1,49 +1,0 @@
-#stdout
-=== Test NewSymbolProcessor ===
-Before: No prefixes After: Prefixes: [lua_ luaL_]
-
-=== Test GenMethodName ===
-Before: Class: INIReader, Name: INIReader After: (*INIReader).Init
-Before: Class: INIReader, Name: INIReader After: (*INIReader).Dispose
-Before: Class: INIReader, Name: HasValue After: (*INIReader).HasValue
-
-=== Test AddSuffix ===
-Before: Class: INIReader, Method: INIReader After: (*Reader).Init
-Before: Class: INIReader, Method: INIReader After: (*Reader).Init__1
-Before: Class: INIReader, Method: ParseError After: (*Reader).ParseError
-Before: Class: INIReader, Method: HasValue After: (*Reader).HasValue
-
-=== Test Case: C++ Class with Methods ===
-Parsed Symbols:
-Symbol Map GoName: (*Reader).Init__1, ProtoName In HeaderFile: INIReader::INIReader(const char *, int), MangledName: _ZN9INIReaderC1EPKci
-Symbol Map GoName: (*Reader).Init, ProtoName In HeaderFile: INIReader::INIReader(const int &), MangledName: _ZN9INIReaderC1ERKi
-Symbol Map GoName: (*Reader).Dispose, ProtoName In HeaderFile: INIReader::~INIReader(), MangledName: _ZN9INIReaderD1Ev
-Symbol Map GoName: (*Reader).ParseError, ProtoName In HeaderFile: INIReader::ParseError(), MangledName: _ZNK9INIReader10ParseErrorEv
-
-=== Test Case: C Functions ===
-Parsed Symbols:
-Symbol Map GoName: (*State).Compare, ProtoName In HeaderFile: lua_compare(lua_State *, int, int, int), MangledName: lua_compare
-Symbol Map GoName: (*State).Rawequal, ProtoName In HeaderFile: lua_rawequal(lua_State *, int, int), MangledName: lua_rawequal
-Symbol Map GoName: Sizecomp, ProtoName In HeaderFile: lua_sizecomp(int, int, int, int), MangledName: lua_sizecomp
-
-=== Test Case: InvalidReceiver ===
-Parsed Symbols:
-Symbol Map GoName: (*Sqlite3).Errcode, ProtoName In HeaderFile: sqlite3_errcode(sqlite3 *), MangledName: sqlite3_errcode
-Symbol Map GoName: UriParameter, ProtoName In HeaderFile: sqlite3_uri_parameter(sqlite3_filename, const char *), MangledName: sqlite3_uri_parameter
-
-=== Test Case: InvalidReceiver PointerLevel > 1 ===
-Parsed Symbols:
-Symbol Map GoName: DerDecoding, ProtoName In HeaderFile: asn1_der_decoding(asn1_node *, const void *, int, char *), MangledName: asn1_der_decoding
-
-=== Test Case: InvalidReceiver typ.NamedType.String is empty ===
-Parsed Symbols:
-Symbol Map GoName: InitWindow, ProtoName In HeaderFile: InitWindow(int, int, const char *), MangledName: InitWindow
-
-=== Test Case: InvalidReceiver typ.canonicalType.Kind == clang.TypePointer ===
-Parsed Symbols:
-Symbol Map GoName: X__mpzSetUiSafe, ProtoName In HeaderFile: __mpz_set_ui_safe(mpz_ptr, unsigned long), MangledName: __mpz_set_ui_safe
-
-
-#stderr
-
-#exit 0
